requirement.txt
diff-match-patch==20121119
Django==1.11.7
django-crispy-forms==1.7.0
django-formtools==2.1
django-import-export==0.5.1
django-pure-pagination==0.3.0
django-reversion==2.0.10
django-simple-captcha==0.5.5
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
jdcal==1.3
mysqlclient==1.3.12
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
Pillow==4.3.0
pytz==2017.3
PyYAML==3.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
